refactor: use PageHeader in Gereric page, move dropdown to its own file
