+++ src/hooks/useGetDashboards.ts
@@ -1,13 +1,12 @@
 import { useEffect } from 'react';
 import { getUsersDashboards } from '../api/dashboard-templates';
-import { useAtomValue, useSetAtom } from 'jotai';
+import { useAtom } from 'jotai';
 import { dashboardsAtom } from '../state/dashboardsAtom';
 import useCurrentUser from './useCurrentUser';
 
 const useGetDashboards = () => {
   const { currentUser } = useCurrentUser();
-  const dashboards = useAtomValue(dashboardsAtom);
-  const setDashboards = useSetAtom(dashboardsAtom);
+  const [dashboards, setDashboards] = useAtom(dashboardsAtom);
 
   const fetchDashboards = async () => {
     try {
